ClickHouseAppender now is separated repository
