--- ClickHouseAppender/src/main/java/com/crypto/service/dao/ClickHouseLogDAO.java
@@ -1,38 +1,0 @@
-package com.crypto.service.dao;
-
-import com.clickhouse.client.*;
-import com.crypto.service.util.ConnectionSettings;
-
-import java.io.ByteArrayInputStream;
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-
-public class ClickHouseLogDAO {
-
-  private final ClickHouseNode server;
-  private final ClickHouseClient client;
-  private final String tableName;
-
-  public ClickHouseLogDAO(String tableName, ConnectionSettings connectionSettings) {
-    try {
-      this.server = ConnectionSettings.initClickHouseConnection(connectionSettings);
-    } catch (IOException e) {
-      throw new RuntimeException(e);
-    }
-    this.client = ClickHouseClient.newInstance(server.getProtocol());
-    this.tableName = tableName;
-  }
-
-  public String getTableName() {
-    return tableName;
-  }
-
-  public void insertLogData(String tsvData) throws ClickHouseException {
-    try (ClickHouseResponse response =
-        client
-            .write(server)
-            .query("INSERT INTO " + tableName)
-            .data(new ByteArrayInputStream(tsvData.getBytes(StandardCharsets.UTF_8)))
-            .executeAndWait()) {}
-  }
-}
